.curlStatus {Rhdf5lib}	R Documentation

Determine whether libcurl and libcrypto were found during package installation

Description

Determine whether libcurl and libcrypto were found during package installation

Usage

.curlStatus()

[Package Rhdf5lib version 1.16.0 Index]